Added PlayerMethods files to store utility functions.
