+++ Projects/CoX/Common/NetStructures/CMakeLists.txt
@@ -4,6 +4,7 @@ ${CMAKE_CURRENT_SOURCE_DIR}/CommonNetStructures.cpp
 ${CMAKE_CURRENT_SOURCE_DIR}/Costume.cpp
 ${CMAKE_CURRENT_SOURCE_DIR}/Powers.cpp
 ${CMAKE_CURRENT_SOURCE_DIR}/Entity.cpp
+${CMAKE_CURRENT_SOURCE_DIR}/PlayerMethods.cpp
 ${CMAKE_CURRENT_SOURCE_DIR}/StateInterpolator.cpp
 )
 SET(target_INCLUDE
@@ -12,6 +13,7 @@ ${CMAKE_CURRENT_SOURCE_DIR}/Powers.h
 ${CMAKE_CURRENT_SOURCE_DIR}/Character.h
 ${CMAKE_CURRENT_SOURCE_DIR}/Costume.h
 ${CMAKE_CURRENT_SOURCE_DIR}/Entity.h
+${CMAKE_CURRENT_SOURCE_DIR}/PlayerMethods.h
 ${CMAKE_CURRENT_SOURCE_DIR}/Contact.h
 ${CMAKE_CURRENT_SOURCE_DIR}/StateInterpolator.h
 )
