+++ Projects/CoX/Common/NetStructures/Character.cpp
@@ -90,18 +90,6 @@ void Character::setName(const QString &val )
         m_name = "EMPTY";
 }
 
-void setTitles(Character &src, bool prefix, QString generic, QString origin, QString special)
-{
-    src.m_has_titles = prefix || !generic.isEmpty() || !origin.isEmpty() || special.isEmpty();
-    if(!src.m_has_titles)
-      return;
-
-    src.m_has_the_prefix = prefix;
-    src.m_titles[0] = generic;
-    src.m_titles[1] = origin;
-    src.m_titles[2] = special;
-}
-
 void Character::sendTray(BitStream &bs) const
 {
     m_trays.serializeto(bs);
@@ -682,20 +670,3 @@ void ClientOptions::init()
     };
 }
 #undef ADD_OPT
-
-void toggleAFK(Character &src, const QString &msg)
-{
-    src.m_afk = !src.m_afk;
-    if(src.m_afk)
-        src.m_afk_msg = msg;
-}
-
-void setXP(Character &src, uint32_t val)
-{
-    src.m_experience_points = val;
-    for(auto const& lvl: src.m_other_attribs.m_ExperienceRequired)
-    {
-        if(val>=lvl && val<lvl+1)
-            setLevel(src, lvl);
-    }
-}
