+++ Projects/CoX/Common/NetStructures/Entity.h
@@ -247,7 +247,6 @@ public:
         void                dump();
         void                addPosUpdate(const PosUpdate &p);
         void                addInterp(const PosUpdate &p);
-        void                toggleFly(Entity *e);
 
 static  void                sendAllyID(BitStream &bs);
 static  void                sendPvP(BitStream &bs);
@@ -262,21 +261,3 @@ static  void                sendPvP(BitStream &bs);
 void initializeNewPlayerEntity(Entity &e);
 void fillEntityFromNewCharData(Entity &e,BitStream &src, ColorAndPartPacker *packer);
 extern void abortLogout(Entity *e);
-
-uint32_t    getIdx(Entity &e) { return e.m_idx; }
-uint32_t    getDbId(Entity &e) { return e.m_db_id; }
-void        setDbId(Entity &e, uint8_t val) { e.m_db_id = val; }
-
-void    toggleFly(Entity &e) { e.m_is_flying = !e.m_is_flying; }
-void    toggleStunned(Entity &e);
-void    toggleJumppack(Entity &e);
-void    setBackupSpd(Entity &e, const float &val) { e.m_backup_spd = val; }
-void    setJumpHeight(Entity &e, const float &val) { e.m_jump_height = val; }
-void    toggleControlsDisabled(Entity &e) { e.m_controls_disabled = !e.m_controls_disabled; }
-void    setUpdateID(Entity &e, const uint8_t &val) { e.m_update_id = val;}
-void    setu1(Entity &e, const int &val) { e.u1 = val; }
-void    setu2(Entity &e, const int &val) { e.u2 = val; }
-void    setu3(Entity &e, const int &val) { e.u3 = val; }
-void    setu4(Entity &e, const int &val) { e.u4 = val; }
-void    setu5(Entity &e, const int &val) { e.u5 = val; }
-void    setu6(Entity &e, const int &val) { e.u6 = val; }
